02-use-cases/ai_coding/requirements.txt
veadk-python==0.2.27
agentkit-sdk-python==0.1.15
tos>=2.8.7
google-adk==1.18.0
